Big commit. Read description
This commit does everything to deal with creating polls and adding new candidates to it along with declaring results for that poll.

Join poll option also edited in order to add joined polls to dashboard.

Made the candidate cards clickable cards to show full description and photo of the candidate in the joined poll section.